refactor(calendar): extract sync + import state machine into CalendarSyncService
